Move MusicBrainzService to Lib.Services
