+++ src/Lib.Services/Extensions/Telemetry/MusicBrainzServiceActivityExtensions.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics;
-using MuzakBot.App.Models.Diagnostics;
+using MuzakBot.Lib.Models.Diagnostics;
 
-namespace MuzakBot.App.Extensions;
+namespace MuzakBot.Lib.Services.Extensions.Telemetry;
 
 /// <summary>
 /// Provides extension methods for the <see cref="ActivitySource"/> class related to the <see cref="MusicBrainzApiService"/>.
